requirements.txt
anyio==3.6.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
ecdsa==0.18.0
fastapi==0.85.0
h11==0.14.0
idna==3.4
pyasn1==0.4.8
pydantic==1.10.2
python-jose==3.3.0
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
